# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Suivicargo API
  description: |
    API de gestion et de suivi de cargaison par fret maritime.

    Authentification : JWT Bearer dans le header `Authorization: Bearer <token>`.
    Le token est obtenu via `POST /v1/auth/login`.

    ## Rôles
    - **ADMIN** : gestion des utilisateurs (employés/superviseurs)
    - **SUPERVISOR** : gestion opérationnelle (navires, voyages, conteneurs, plans de paiement, statuts)
    - **EMPLOYEE** : enlèvement, encaissement, lecture des cargaisons en cours
  version: 1.0.0
  contact:
    name: Equipe Suivicargo
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api
    description: Local development
  - url: https://api.suivicargo.example.com
    description: Production
...
security:
  - bearerAuth: []

# =====================================================================
# Paths
# =====================================================================
paths:
...
  /v1/users:
    get:
      tags: [Users]
      summary: Lister les utilisateurs (ADMIN)
      operationId: listUsers
      parameters:
        - in: query
          name: role
          schema:
            $ref: '#/components/schemas/Role'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page d'utilisateurs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseUser'
        '403':
          $ref: '#/components/responses/Forbidden'
  /v1/users/{id}:
    get:
      tags: [Users]
      summary: Détail utilisateur (ADMIN)
      operationId: getUserById
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Détail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    patch:
      tags: [Users]
      summary: Modifier un utilisateur (ADMIN)
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
    delete:
      tags: [Users]
      summary: Supprimer (soft) un utilisateur (ADMIN)
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Supprimé

  # ---------- Navires ----------
  /v1/navires:
    get:
      tags: [Navires]
      summary: Lister les navires
      operationId: listNavires
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page de navires
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseNavire'
    post:
      tags: [Navires]
      summary: Créer un navire (SUPERVISOR/ADMIN)
      operationId: createNavire
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNavireRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavireDto'
  /v1/navires/{id}:
    get:
      tags: [Navires]
      summary: Détail navire
      operationId: getNavireById
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Détail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavireDto'
    patch:
      tags: [Navires]
      summary: Modifier un navire
      operationId: updateNavire
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNavireRequest'
      responses:
        '200':
          description: Modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NavireDto'
    delete:
      tags: [Navires]
      summary: Supprimer un navire
      operationId: deleteNavire
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Supprimé

  # ---------- Voyages ----------
  /v1/voyages:
    get:
      tags: [Voyages]
      summary: Lister les voyages
      operationId: listVoyages
      parameters:
        - in: query
          name: navireId
          schema: { type: integer, format: int64 }
        - in: query
          name: statut
          schema:
            $ref: '#/components/schemas/StatutVoyage'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page de voyages
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseVoyage'
    post:
      tags: [Voyages]
      summary: Créer un voyage (SUPERVISOR/ADMIN)
      operationId: createVoyage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVoyageRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoyageDto'
  /v1/voyages/{id}:
    get:
      tags: [Voyages]
      operationId: getVoyageById
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '200':
          description: Détail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoyageDto'
    patch:
      tags: [Voyages]
      operationId: updateVoyage
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateVoyageRequest'
      responses:
        '200':
          description: Modifié
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoyageDto'
    delete:
      tags: [Voyages]
      operationId: deleteVoyage
      parameters: [{ $ref: '#/components/parameters/IdParam' }]
      responses:
        '204':
          description: Supprimé

  # ---------- Conteneurs ----------
...
  /v1/plans-paiement:
    post:
      tags: [Plans de paiement]
      summary: Créer un plan de paiement (SUPERVISOR/ADMIN)
      operationId: createPlanPaiement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePlanPaiementRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanPaiementDto'
  /v1/plans-paiement/by-cargaison/{cargaisonId}:
    get:
      tags: [Plans de paiement]
      summary: Plan de paiement d'une cargaison
      operationId: getPlanByCargaison
      parameters:
        - in: path
          name: cargaisonId
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Détail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanPaiementDto'
        '404':
          $ref: '#/components/responses/NotFound'

  # ---------- Règlements ----------
  /v1/reglements:
    post:
      tags: [Règlements]
      summary: Enregistrer un règlement
      operationId: createReglement
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReglementRequest'
      responses:
        '201':
          description: Créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReglementDto'
  /v1/reglements/by-cargaison/{cargaisonId}:
    get:
      tags: [Règlements]
      summary: Règlements d'une cargaison
      operationId: listReglementsByCargaison
      parameters:
        - in: path
          name: cargaisonId
          required: true
          schema: { type: integer, format: int64 }
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
      responses:
        '200':
          description: Page de règlements
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageResponseReglement'

  # ---------- Factures ----------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu via `POST /v1/auth/login`

  # ---------- Reusable parameters ----------
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
    PageParam:
      in: query
      name: page
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Numéro de page (0-indexé)
    SizeParam:
      in: query
      name: size
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 20
      description: Nombre d'éléments par page
...
  responses:
...
    Forbidden:
      description: Authentifié mais sans le rôle requis
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
...
  schemas:
...
    Role:
      type: string
      enum: [ADMIN, SUPERVISOR, EMPLOYEE]
...
    StatutPaiement:
      type: string
      enum: [EN_ATTENTE, PARTIEL, PAYE, EN_RETARD, ANNULE]
    StatutVoyage:
      type: string
      enum: [PROGRAMME, EN_MER, ARRIVE, ANNULE]
    ModePaiement:
      type: string
      enum: [ESPECES, VIREMENT, MOBILE_MONEY, CHEQUE, CARTE]

    # --- Auth ---
...
    ApiError:
      type: object
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: Bad Request
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        path:
          type: string
        fieldErrors:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/FieldErrorDetail'
    FieldErrorDetail:
      type: object
      required: [field, message]
      properties:
        field:
          type: string
        message:
          type: string
        rejectedValue:
          nullable: true

    # --- Client ---
...
    UserDto:
      type: object
      required: [id, email, nom, prenom, role, actif]
      properties:
        id: { type: integer, format: int64 }
        email: { type: string, format: email }
        nom: { type: string }
        prenom: { type: string }
        telephone: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/Role' }
        actif: { type: boolean }
        dateCreation: { type: string, format: date-time }
    UpdateUserRequest:
      type: object
      properties:
        nom: { type: string, maxLength: 100 }
        prenom: { type: string, maxLength: 100 }
        telephone: { type: string, maxLength: 30 }
        role: { $ref: '#/components/schemas/Role' }
        actif: { type: boolean }
    PageResponseUser:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/UserDto' }

    # --- Navire ---
    NavireDto:
      type: object
      required: [id, nom, imo]
      properties:
        id: { type: integer, format: int64 }
        nom: { type: string }
        imo: { type: string, pattern: '^\d{7}$' }
        pavillon: { type: string, nullable: true }
        capaciteEvp: { type: integer, nullable: true }
    CreateNavireRequest:
      type: object
      required: [nom, imo]
      properties:
        nom: { type: string, maxLength: 150 }
        imo: { type: string, pattern: '^\d{7}$' }
        pavillon: { type: string, maxLength: 50 }
        capaciteEvp: { type: integer, minimum: 1 }
    UpdateNavireRequest:
      type: object
      properties:
        nom: { type: string, maxLength: 150 }
        pavillon: { type: string, maxLength: 50 }
        capaciteEvp: { type: integer, minimum: 1 }
    PageResponseNavire:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/NavireDto' }

    # --- Voyage ---
    VoyageDto:
      type: object
      required: [id, navireId, portDepart, portArrivee, dateDepart, etaArrivee, statut]
      properties:
        id: { type: integer, format: int64 }
        navireId: { type: integer, format: int64 }
        navireNom: { type: string }
        portDepart: { type: string }
        portArrivee: { type: string }
        dateDepart: { type: string, format: date }
        etaArrivee: { type: string, format: date }
        dateArriveeReelle: { type: string, format: date, nullable: true }
        statut: { $ref: '#/components/schemas/StatutVoyage' }
    CreateVoyageRequest:
      type: object
      required: [navireId, portDepart, portArrivee, dateDepart, etaArrivee]
      properties:
        navireId: { type: integer, format: int64 }
        portDepart: { type: string, maxLength: 100 }
        portArrivee: { type: string, maxLength: 100 }
        dateDepart: { type: string, format: date }
        etaArrivee: { type: string, format: date }
    UpdateVoyageRequest:
      type: object
      properties:
        portDepart: { type: string, maxLength: 100 }
        portArrivee: { type: string, maxLength: 100 }
        dateDepart: { type: string, format: date }
        etaArrivee: { type: string, format: date }
        dateArriveeReelle: { type: string, format: date }
        statut: { $ref: '#/components/schemas/StatutVoyage' }
    PageResponseVoyage:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/VoyageDto' }

    # --- Conteneur ---
...
    EcheanceDto:
      type: object
      required: [id, ordre, libelle, montantPrevu, dateEcheance, statut]
      properties:
        id: { type: integer, format: int64 }
        ordre: { type: integer }
        libelle: { type: string }
        montantPrevu: { type: number, format: double }
        dateEcheance: { type: string, format: date }
        statut: { $ref: '#/components/schemas/StatutPaiement' }
    CreateEcheanceRequest:
      type: object
      required: [ordre, libelle, montantPrevu, dateEcheance]
      properties:
        ordre: { type: integer, minimum: 1 }
        libelle: { type: string, maxLength: 100 }
        montantPrevu: { type: number, format: double, minimum: 0 }
        dateEcheance: { type: string, format: date }
    PlanPaiementDto:
      type: object
      required: [id, cargaisonId, montantTotal, statut, echeances]
      properties:
        id: { type: integer, format: int64 }
        cargaisonId: { type: integer, format: int64 }
        numeroTracage: { type: string }
        montantTotal: { type: number, format: double }
        montantRegle: { type: number, format: double }
        montantRestant: { type: number, format: double }
        devise: { type: string }
        statut: { $ref: '#/components/schemas/StatutPaiement' }
        echeances:
          type: array
          items: { $ref: '#/components/schemas/EcheanceDto' }
    CreatePlanPaiementRequest:
      type: object
      required: [cargaisonId, echeances]
      properties:
        cargaisonId: { type: integer, format: int64 }
        devise: { type: string, minLength: 3, maxLength: 3 }
        echeances:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/CreateEcheanceRequest' }

    # --- Reglement ---
    ReglementDto:
      type: object
      required: [id, montant, modePaiement, dateReglement]
      properties:
        id: { type: integer, format: int64 }
        planPaiementId: { type: integer, format: int64, nullable: true }
        echeanceId: { type: integer, format: int64, nullable: true }
        montant: { type: number, format: double }
        modePaiement: { $ref: '#/components/schemas/ModePaiement' }
        referenceTransaction: { type: string, nullable: true }
        dateReglement: { type: string, format: date }
        encaissePaiUserId: { type: integer, format: int64, nullable: true }
        encaissePar: { type: string, nullable: true }
        commentaire: { type: string, nullable: true }
    CreateReglementRequest:
      type: object
      required: [cargaisonId, montant, modePaiement, dateReglement]
      properties:
        cargaisonId: { type: integer, format: int64 }
        montant: { type: number, format: double, minimum: 0 }
        modePaiement: { $ref: '#/components/schemas/ModePaiement' }
        referenceTransaction: { type: string, maxLength: 100 }
        dateReglement: { type: string, format: date }
        echeanceId: { type: integer, format: int64, nullable: true }
        commentaire: { type: string, maxLength: 500 }
    PageResponseReglement:
      allOf:
        - $ref: '#/components/schemas/PageBase'
        - type: object
          properties:
            content:
              type: array
              items: { $ref: '#/components/schemas/ReglementDto' }

    # --- Facture ---
...
    PageBase:
      type: object
      required: [page, size, totalElements, totalPages, first, last]
      properties:
        page: { type: integer }
        size: { type: integer }
        totalElements: { type: integer, format: int64 }
        totalPages: { type: integer }
        first: { type: boolean }
        last: { type: boolean }